+++ mayan/apps/fundraisers/literals.py
@@ -1,1 +1,1 @@
-URL_FUNDRAISER_API_LOGIN = 'https://fundraiser.mayan-edms.com/locations/login'
+URL_FUNDRAISER_API = 'https://fundraiser.mayan-edms.com'